requirements.txt
starlette>=0.37.2
pydantic>=2
rich>=13
pytest>=8
numpy>=1
setuptools>=68
requests>=2
properscoring>=0.1
pandas>=2.2.3
bittensor>=8.4.3
alembic>=1.14.0
python-dotenv>=1.0.1
psycopg2-binary>=2.9.10
sqlalchemy>=2.0.36
testcontainers>=4.9.0
